Disable condition for non static metric alarm
